Break up main.go
Add new files for abspath, credential, and env functions.
