+++ main.go
@@ -21,7 +21,6 @@ package main // import "k8s.io/git-sync/cmd/git-sync"
 import (
 	"context"
 	"crypto/md5"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
@@ -105,362 +104,6 @@ const (
 )
 
 const defaultDirMode = os.FileMode(0775) // subject to umask
-
-type credential struct {
-	URL          string `json:"url"`
-	Username     string `json:"username"`
-	Password     string `json:"password,omitempty"`
-	PasswordFile string `json:"password-file,omitempty"`
-}
-
-func (c credential) String() string {
-	jb, err := json.Marshal(c)
-	if err != nil {
-		return fmt.Sprintf("<encoding error: %v>", err)
-	}
-	return string(jb)
-}
-
-// credentialSliceValue is for flags.
-type credentialSliceValue struct {
-	value   []credential
-	changed bool
-}
-
-var _ pflag.Value = &credentialSliceValue{}
-var _ pflag.SliceValue = &credentialSliceValue{}
-
-// pflagCredentialSlice is like pflag.StringSlice()
-func pflagCredentialSlice(name, def, usage string) *[]credential {
-	p := &credentialSliceValue{}
-	_ = p.Set(def)
-	pflag.Var(p, name, usage)
-	return &p.value
-}
-
-// unmarshal is like json.Unmarshal, but fails on unknown fields.
-func (cs credentialSliceValue) unmarshal(val string, out any) error {
-	dec := json.NewDecoder(strings.NewReader(val))
-	dec.DisallowUnknownFields()
-	return dec.Decode(out)
-}
-
-// decodeList handles a string-encoded JSON object.
-func (cs credentialSliceValue) decodeObject(val string) (credential, error) {
-	var cred credential
-	if err := cs.unmarshal(val, &cred); err != nil {
-		return credential{}, err
-	}
-	return cred, nil
-}
-
-// decodeList handles a string-encoded JSON list.
-func (cs credentialSliceValue) decodeList(val string) ([]credential, error) {
-	var creds []credential
-	if err := cs.unmarshal(val, &creds); err != nil {
-		return nil, err
-	}
-	return creds, nil
-}
-
-// decode handles a string-encoded JSON object or list.
-func (cs credentialSliceValue) decode(val string) ([]credential, error) {
-	s := strings.TrimSpace(val)
-	if s == "" {
-		return nil, nil
-	}
-	// If it tastes like an object...
-	if s[0] == '{' {
-		cred, err := cs.decodeObject(s)
-		return []credential{cred}, err
-	}
-	// If it tastes like a list...
-	if s[0] == '[' {
-		return cs.decodeList(s)
-	}
-	// Otherwise, bad
-	return nil, fmt.Errorf("not a JSON object or list")
-}
-
-func (cs *credentialSliceValue) Set(val string) error {
-	v, err := cs.decode(val)
-	if err != nil {
-		return err
-	}
-
-	if !cs.changed {
-		cs.value = v
-	} else {
-		cs.value = append(cs.value, v...)
-	}
-	cs.changed = true
-
-	return nil
-}
-
-func (cs credentialSliceValue) Type() string {
-	return "credentialSlice"
-}
-
-func (cs credentialSliceValue) String() string {
-	if len(cs.value) == 0 {
-		return "[]"
-	}
-	jb, err := json.Marshal(cs.value)
-	if err != nil {
-		return fmt.Sprintf("<encoding error: %v>", err)
-	}
-	return string(jb)
-}
-
-func (cs *credentialSliceValue) Append(val string) error {
-	v, err := cs.decodeObject(val)
-	if err != nil {
-		return err
-	}
-	cs.value = append(cs.value, v)
-	return nil
-}
-
-func (cs *credentialSliceValue) Replace(val []string) error {
-	creds := []credential{}
-	for _, s := range val {
-		v, err := cs.decodeObject(s)
-		if err != nil {
-			return err
-		}
-		creds = append(creds, v)
-	}
-	cs.value = creds
-	return nil
-}
-
-func (cs credentialSliceValue) GetSlice() []string {
-	if len(cs.value) == 0 {
-		return nil
-	}
-	ret := []string{}
-	for _, cred := range cs.value {
-		ret = append(ret, cred.String())
-	}
-	return ret
-}
-
-func envString(def string, key string, alts ...string) string {
-	if val := os.Getenv(key); val != "" {
-		return val
-	}
-	for _, alt := range alts {
-		if val := os.Getenv(alt); val != "" {
-			fmt.Fprintf(os.Stderr, "env %s has been deprecated, use %s instead\n", alt, key)
-			return val
-		}
-	}
-	return def
-}
-
-func envStringArray(def string, key string, alts ...string) []string {
-	parse := func(s string) []string {
-		return strings.Split(s, ":")
-	}
-
-	if val := os.Getenv(key); val != "" {
-		return parse(val)
-	}
-	for _, alt := range alts {
-		if val := os.Getenv(alt); val != "" {
-			fmt.Fprintf(os.Stderr, "env %s has been deprecated, use %s instead\n", alt, key)
-			return parse(val)
-		}
-	}
-	return parse(def)
-}
-
-func envBoolOrError(def bool, key string, alts ...string) (bool, error) {
-	parse := func(key, val string) (bool, error) {
-		parsed, err := strconv.ParseBool(val)
-		if err == nil {
-			return parsed, nil
-		}
-		return false, fmt.Errorf("ERROR: invalid bool env %s=%q: %w", key, val, err)
-	}
-
-	if val := os.Getenv(key); val != "" {
-		return parse(key, val)
-	}
-	for _, alt := range alts {
-		if val := os.Getenv(alt); val != "" {
-			fmt.Fprintf(os.Stderr, "env %s has been deprecated, use %s instead\n", alt, key)
-			return parse(alt, val)
-		}
-	}
-	return def, nil
-}
-func envBool(def bool, key string, alts ...string) bool {
-	val, err := envBoolOrError(def, key, alts...)
-	if err != nil {
-		fmt.Fprintln(os.Stderr, err)
-		os.Exit(1)
-		return false
-	}
-	return val
-}
-
-func envIntOrError(def int, key string, alts ...string) (int, error) {
-	parse := func(key, val string) (int, error) {
-		parsed, err := strconv.ParseInt(val, 0, 0)
-		if err == nil {
-			return int(parsed), nil
-		}
-		return 0, fmt.Errorf("ERROR: invalid int env %s=%q: %w", key, val, err)
-	}
-
-	if val := os.Getenv(key); val != "" {
-		return parse(key, val)
-	}
-	for _, alt := range alts {
-		if val := os.Getenv(alt); val != "" {
-			fmt.Fprintf(os.Stderr, "env %s has been deprecated, use %s instead\n", alt, key)
-			return parse(alt, val)
-		}
-	}
-	return def, nil
-}
-func envInt(def int, key string, alts ...string) int {
-	val, err := envIntOrError(def, key, alts...)
-	if err != nil {
-		fmt.Fprintln(os.Stderr, err)
-		os.Exit(1)
-		return 0
-	}
-	return val
-}
-
-func envFloatOrError(def float64, key string, alts ...string) (float64, error) {
-	parse := func(key, val string) (float64, error) {
-		parsed, err := strconv.ParseFloat(val, 64)
-		if err == nil {
-			return parsed, nil
-		}
-		return 0, fmt.Errorf("ERROR: invalid float env %s=%q: %w", key, val, err)
-	}
-
-	if val := os.Getenv(key); val != "" {
-		return parse(key, val)
-	}
-	for _, alt := range alts {
-		if val := os.Getenv(alt); val != "" {
-			fmt.Fprintf(os.Stderr, "env %s has been deprecated, use %s instead\n", alt, key)
-			return parse(alt, val)
-		}
-	}
-	return def, nil
-}
-func envFloat(def float64, key string, alts ...string) float64 {
-	val, err := envFloatOrError(def, key, alts...)
-	if err != nil {
-		fmt.Fprintln(os.Stderr, err)
-		os.Exit(1)
-		return 0
-	}
-	return val
-}
-
-func envDurationOrError(def time.Duration, key string, alts ...string) (time.Duration, error) {
-	parse := func(key, val string) (time.Duration, error) {
-		parsed, err := time.ParseDuration(val)
-		if err == nil {
-			return parsed, nil
-		}
-		return 0, fmt.Errorf("ERROR: invalid duration env %s=%q: %w", key, val, err)
-	}
-
-	if val := os.Getenv(key); val != "" {
-		return parse(key, val)
-	}
-	for _, alt := range alts {
-		if val := os.Getenv(alt); val != "" {
-			fmt.Fprintf(os.Stderr, "env %s has been deprecated, use %s instead\n", alt, key)
-			return parse(alt, val)
-		}
-	}
-	return def, nil
-}
-func envDuration(def time.Duration, key string, alts ...string) time.Duration {
-	val, err := envDurationOrError(def, key, alts...)
-	if err != nil {
-		fmt.Fprintln(os.Stderr, err)
-		os.Exit(1)
-		return 0
-	}
-	return val
-}
-
-// absPath is an absolute path string.  This type is intended to make it clear
-// when strings are absolute paths vs something else.  This does not verify or
-// mutate the input, so careless callers could make instances of this type that
-// are not actually absolute paths, or even "".
-type absPath string
-
-// String returns abs as a string.
-func (abs absPath) String() string {
-	return string(abs)
-}
-
-// Canonical returns a canonicalized form of abs, similar to filepath.Abs
-// (including filepath.Clean).  Unlike filepath.Clean, this preserves "" as a
-// special case.
-func (abs absPath) Canonical() (absPath, error) {
-	if abs == "" {
-		return abs, nil
-	}
-
-	result, err := filepath.Abs(abs.String())
-	if err != nil {
-		return "", err
-	}
-	return absPath(result), nil
-}
-
-// Join appends more path elements to abs, like filepath.Join.
-func (abs absPath) Join(elems ...string) absPath {
-	all := make([]string, 0, 1+len(elems))
-	all = append(all, abs.String())
-	all = append(all, elems...)
-	return absPath(filepath.Join(all...))
-}
-
-// Split breaks abs into stem and leaf parts (often directory and file, but not
-// necessarily), similar to filepath.Split.  Unlike filepath.Split, the
-// resulting stem part does not have any trailing path separators.
-func (abs absPath) Split() (string, string) {
-	if abs == "" {
-		return "", ""
-	}
-
-	// filepath.Split promises that dir+base == input, but trailing slashes on
-	// the dir is confusing and ugly.
-	pathSep := string(os.PathSeparator)
-	dir, base := filepath.Split(strings.TrimRight(abs.String(), pathSep))
-	dir = strings.TrimRight(dir, pathSep)
-	if len(dir) == 0 {
-		dir = string(os.PathSeparator)
-	}
-
-	return dir, base
-}
-
-// Dir returns the stem part of abs without the leaf, like filepath.Dir.
-func (abs absPath) Dir() string {
-	dir, _ := abs.Split()
-	return dir
-}
-
-// Base returns the leaf part of abs without the stem, like filepath.Base.
-func (abs absPath) Base() string {
-	_, base := abs.Split()
-	return base
-}
 
 // repoSync represents the remote repo and the local sync of it.
 type repoSync struct {
